reorganized src folder, updated imports to match, substituted eslint for tslint as a dependency
