+++ package.json
@@ -18,6 +18,7 @@
     "@types/redux-logger": "^3.0.8",
     "autoprefixer": "^9",
     "classnames": "2.2.6",
+    "eslint": "^7.20.0",
     "firebase": "^8.2.2",
     "headroom.js": "0.11.0",
     "moment": "2.26.0",
