+++ src/assets/App.scss
@@ -1,4 +1,4 @@
-@import "./assets/styles.scss";
+@import "styles.scss";
 
 .App {
   text-align: center;
