+++ src/components/ContactUs.tsx
@@ -7,7 +7,7 @@
  */
 import { useState } from "react";
 import { Button, Form, Input, Alert } from "reactstrap";
-import { createEmailDocument } from "../FireConfig";
+import { createEmailDocument } from "../scripts/FireConfig";
 import { Redirect } from "react-router-dom";
 import "../assets/styles.scss";
