+++ src/scripts/FireConfig.ts
